requirements.txt
Flask
Flask-SocketIO
pyserial
Flask-Cors